quic: refactor crypto in prep of ngtcp2_crypto
PR-URL: https://github.com/nodejs/quic/pull/104
