+++ node.gyp
@@ -835,6 +835,7 @@
             'src/node_quic_stream.h',
             'src/node_quic_util.h',
             'src/node_quic_state.h',
+            'src/node_quic_crypto.cc',
             'src/node_quic_session.cc',
             'src/node_quic_socket.cc',
             'src/node_quic_stream.cc',
